+++ download.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+curl -L https://github.com/ArkEcosystem/peers/raw/master/mainnet.json -o ark-mainnet.json
+curl -L https://github.com/ArkEcosystem/peers/raw/master/devnet.json -o ark-devnet.json
+curl -L https://github.com/Internet-of-People/hydra-peers/raw/master/mainnet.json -o hyd-mainnet.json
+curl -L https://github.com/Internet-of-People/hydra-peers/raw/master/devnet.json -o hyd-devnet.json