+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [2.0.0] - 2020-04-07
 
-- Support for Hydra peers
+- Support for Hydra peers and network definition files.
+- Storing some extra information about network nodes in a
+  timestamped output file.
 
 ## [1.3.0] - 2020-02-04